Refactoring methods doc file and fix tests
