+++ lib/inquisition/outputter/doc.rb
@@ -1,5 +1,6 @@
 require_relative 'html/template'
-require_relative 'documentation/doc_view_helper'
+require_relative 'docx/doc_view_helper'
+require_relative 'docx/doc_file_helper'
 
 module Inquisition
   module Outputter
@@ -8,31 +9,10 @@ module Inquisition
 
       def initialize(_output); end
 
-      def stop(issues)
-        generate_doc(issues)
-      end
-
-      def generate_doc(_issues)
-        documentation = HTML::Template.new('documentation_word').render(Documentation::DocViewHelper.new)
-        File.open(path_to_word_file, 'w') { |file| file.write(documentation) }
-      end
-
-      private
-
-      def path_to_word_file
-        Configuration.instance.output_path + define_file_name
-      end
-
-      def define_file_name
-        "/#{project_file_name}_#{date_file_name}.docx"
-      end
-
-      def project_file_name
-        Rails.application.class.parent_name
-      end
-
-      def date_file_name
-        Time.now.strftime('%d_%m_%Y')
+      def stop(_issues)
+        file_helper = Docx::DocFileHelper.new
+        documentation = HTML::Template.new('documentation_word').render(Docx::DocViewHelper.new)
+        File.open(file_helper.path_to_word_file, 'w') { |file| file.write(documentation) }
       end
     end
   end
